move files out of components folder in insurance
