+++ client/packages/system/src/Patient/Insurance/InsuranceProvidersSelect.tsx
@@ -3,7 +3,7 @@ import React, { FC, ReactElement } from 'react';
 import { useTranslation } from '@common/intl';
 import { Autocomplete, InputWithLabelRow } from '@common/components';
 
-import { usePatient } from '../../../api';
+import { usePatient } from '../api';
 
 interface InsuranceProvidersSelectProps {
   insuranceProviderId: string;
