--- client/packages/system/src/Patient/Insurance/components/InsuranceModal/index.ts
@@ -1,1 +1,0 @@
-export * from './InsuranceModal';
