--- client/packages/system/src/Patient/Insurance/components/InsurancePolicySelect/index.ts
@@ -1,1 +1,0 @@
-export * from './InsurancePolicySelect';
