--- client/packages/system/src/Patient/Insurance/components/InsuranceProvidersSelect/index.ts
@@ -1,1 +1,0 @@
-export * from './InsuranceProvidersSelect';
