--- client/packages/system/src/Patient/Insurance/components/InsuranceStatusToggle/index.ts
@@ -1,1 +1,0 @@
-export * from './InsuranceStatusToggle';
