--- client/packages/system/src/Patient/Insurance/components/index.ts
@@ -1,4 +1,0 @@
-export * from './InsuranceModal';
-export * from './InsurancePolicySelect';
-export * from './InsuranceProvidersSelect';
-export * from './InsuranceStatusToggle';
