+++ client/packages/system/src/Patient/Insurance/index.ts
@@ -1,2 +1,2 @@
-export * from './components';
+export * from './InsuranceModal';
 export * from './ListView';
